Update ReadInput.java
There is currently a bug in scanner that won't recognize the first char of a string. I have tried the scanner, creating a char array using .getChars and I have tried just a simple for loop to manually create the char array...

I will come back to this and this is the closest to working code I have.